fix(cleanup): cleanup
